requirements.txt
flask==2.3.3
line-bot-sdk==3.5.0
google-generativeai==0.3.2
gunicorn==21.2.0